# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quant Agent Data Gateway",
    "description": "Quant Agent 执行引擎 HTTP/WS/SSE 契约（BE-19）。\n\n## 统一响应\n\n```json\n{\"code\": 0, \"msg\": \"ok\", \"data\": {}, \"ts\": 1719475200000}\n```\n\n- `code=0` 成功；非零见错误码（`docs/10` §1.4）\n- 响应头 `X-Trace-Id` 用于全链路追踪（BE-10）\n\n## 机器可读契约\n\n导出产物：`docs/openapi.json`（`python scripts/export_openapi.py`）。\n人工互校文档：`docs/10. API接口规范.md`。",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/settings/preferences": {
      "get": {
        "tags": [
          "Preferences"
        ],
        "summary": "Get Preferences",
        "description": "获取当前登录用户的全局偏好设置（含绑定的各类 API Keys）",
        "operationId": "get_preferences_api_v1_settings_preferences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Preferences"
        ],
        "summary": "Update Preferences",
        "description": "更新当前登录用户的偏好设置与 API Keys",
        "operationId": "update_preferences_api_v1_settings_preferences_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Prefs"
              },
              "example": {
                "_note": "见 Schema 字段说明；统一响应见 {code,msg,data,ts}"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/news-tags": {
      "get": {
        "tags": [
          "Preferences"
        ],
        "summary": "Get News Tags Rules",
        "description": "获取动态新闻打标规则",
        "operationId": "get_news_tags_rules_api_v1_settings_news_tags_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Preferences"
        ],
        "summary": "Update News Tags Rules",
        "description": "更新动态新闻打标规则",
        "operationId": "update_news_tags_rules_api_v1_settings_news_tags_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object",
                "title": "Rules"
              },
              "example": {
                "_note": "见 Schema 字段说明；统一响应见 {code,msg,data,ts}"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/watchlist": {
      "get": {
        "tags": [
          "Preferences"
        ],
        "summary": "Get Watchlist",
        "description": "获取当前用户的长期监控池股票列表",
        "operationId": "get_watchlist_api_v1_settings_watchlist_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/watchlist/batch": {
      "post": {
        "tags": [
          "Preferences"
        ],
        "summary": "Batch Update Watchlist",
        "description": "一键批量将股票加入或移出系统级监控池，联动后台守护进程",
        "operationId": "batch_update_watchlist_api_v1_settings_watchlist_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistBatchRequest"
              },
              "example": {
                "_note": "见 Schema 字段说明；统一响应见 {code,msg,data,ts}"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WatchlistBatchRequest": {
        "properties": {
          "tickers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tickers"
          },
          "action": {
            "type": "string",
            "title": "Action",
            "default": "add"
          }
        },
        "type": "object",
        "required": [
          "tickers"
        ],
        "title": "WatchlistBatchRequest"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}